Reimplemented Command Loop and Runner
The old command loop was just kind of wedged together.  The new command loop is
built around the idea of trying a command at a stage, then handling the result.
This separate handler allows the code to use return values, which lets it be
smart about starting new VMs.  This addresses the old, ugly behavior which
caused a VM to be started if there was a command that didn't exist.

This required significant changes.  Delegates were reworked.  Delegate commands
were made synchronous (although, without timeouts) to help prevent issues with
EXITs outrunning commands.

Finally, to insulate the command loop from the issues of networking, all
commands are spawned as new processed (even local ones) that communicate via an
intermediary.  This intermediary traps exits and detects networking errors, so
the command loop can be naive.  The remote process is set up such that the
delegate is available, missing commands are detected, and other exceptions are
caught.  The intermediary takes the return value (or exception or lack of
command) and synthesizes an appropriate return value.

Put together, the new command loop is smarter and more resilient to failure.
At least, I hope.  :)